delete old OpenGL renderer code
