+++ Cargo.toml
@@ -4,11 +4,7 @@ members = [
     "ldraw",
     "olr",
     "renderer",
-    "renderer_v2",
     "tools/baker",
     "tools/ldr2img",
-    "tools/viewer/common",
-    "tools/viewer/native",
-    "tools/viewer/web",
     "tools/viewer_v2"
 ]
