--- renderer/shaders/edge.fs
@@ -1,9 +1,0 @@
-precision mediump float;
-
-in vec4 vColor;
-
-out vec4 fragColor;
-
-void main(void) {
-     fragColor = vColor;
-}
