+++ renderer/shaders/model_fragment_base.wgsl
