+++ renderer/shaders/model_fragment_no_shading.wgsl
