+++ renderer/shaders/model_vertex.wgsl
