--- renderer/shaders/optional_edge.fs
@@ -1,14 +1,0 @@
-precision mediump float;
-
-in vec4 vColor;
-in float discardFlag;
-
-out vec4 fragColor;
-
-void main(void) {
-     if (discardFlag > 0.5f) {
-          discard;
-     }
-
-     fragColor = vColor;
-}
