+++ renderer/shaders/optional_edge_vertex.wgsl
