+++ renderer/src/lib.rs
@@ -1,7 +1,4 @@
+pub mod camera;
 pub mod display_list;
-pub mod error;
-pub mod model;
 pub mod part;
-pub mod shader;
-pub mod state;
-pub mod utils;
+pub mod pipeline;
