--- renderer_v2/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "ldraw-renderer-wgpu"
-version = "0.0.1"
-authors = ["Park Joon-Kyu <segfault87@gmail.com>"]
-edition = "2021"
-
-[dependencies]
-bytemuck = { version = "1.13", features = ["derive"] }
-cgmath = "~0.18.0"
-image = "~0.23.14"
-itertools = "~0.10"
-ldraw = { path = "../ldraw" }
-ldraw-ir = { path = "../ir" }
-serde = { version = "1.0", features = ["derive"] }
-uuid = { version = "1", features = ["v4", "serde", "js"] }
-wgpu = "0.17.0"
