--- renderer_v2/src/lib.rs
@@ -1,4 +1,0 @@
-pub mod camera;
-pub mod display_list;
-pub mod part;
-pub mod pipeline;
